New function "FindSolidIntersections". Much faster than IsSolidInSolid, which will now call this function instead. Still needs more debug.
FindSolidIntersections should lay the groundwork for determining whether
solids are inside other solids, as well as, Joining and Subtracting
solids from one another. It is based on similar concepts to the
triangulation function (plane list instead of line list, odd number of
intercepts, ect...). I wrote this because I couldn't find any methods
and this should be close to O(nlogn) .